Remove readr from dependencies
The usage seems entirely redundant. You are taking a character vector
and passing it as literal data to `read_lines()` which returns the same
character vector as the input.

The current example breaks with readr 2.0.0 as literal data needs to be
denoted with `I()` in readr 2.0.0, but as far as I can tell readr should
just be removed entirely.

In addition the genius website layout seems to have changed, the lyrics
are no longer being retrieved with either the CRAN version or this
change.